requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
asyncpg==0.29.0
pydantic-settings==2.1.0
httpx==0.26.0
